+++ Assets/Scripts/Dungen/DungenManager.cs
@@ -16,7 +16,8 @@ public class DungenManager : MonoBehaviour
 
     public IEnumerator MoveCameraCoroutine(Vector3 TargetLocation)
     {
-        while(Mathf.Pow(TargetLocation.x, 2) + Mathf.Pow(TargetLocation.z, 2) > 0.1)
+        Vector3 toMoveCheck = TargetLocation - m_DungenCam.transform.position;
+        while(Mathf.Pow(toMoveCheck.x, 2) + Mathf.Pow(toMoveCheck.z, 2) > 0.4)
         {
             Vector3 toMove = TargetLocation - m_DungenCam.transform.position;
             m_CameraRB.velocity = toMove.normalized * m_cameraSpeed;
