--- Assets/Scripts/DungenDoor.cs
@@ -1,23 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class DungenDoor : MonoBehaviour
-{
-    [SerializeField] private GameObject m_toRoom;
-    private DungenManager m_dungenManager;
-
-    private void Awake()
-    {
-        m_dungenManager = GameObject.FindObjectOfType<DungenManager>();
-    }
-
-    private void OnTriggerEnter(Collider other)
-    {
-        if(other.tag == "Player")
-        {
-            //StartCoroutine(m_dungenManager.MoveCameraCoroutine(m_toRoom.Ge))
-        }
-    }
-}
-
